Extract standard full OTAs
We don't check hashes or signatures yet, but with manual testing
the extracted content has the same hashes as other tools.

This really extracts in parallel on Unixes only for now,
Rust doesn't have a standard portable interface to write to a
file in parallel, didn't bother implementing for Win yet.

And this only supports the newer format v2 and most common
install operations for full OTAs so far.